+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 The format of this change log follows the advice given at [Keep a CHANGELOG](https://keepachangelog.com).
 
 ## [Unreleased]
+### Fixed
+- The `moodle.Commenting.VariableComment` sniff now accepts the following tags as valid on a `@var`:
+  - `@since`
+  - `@link`
+  - `@deprecated`
 ## [v3.5.0] - 2025-07-17
 ### Changed
 - Bumped dependencies for: